Optimized KvStore Peer Sync
Summary:
Now when KvStore sync with peers (new link established, or randomly select one peer for synchronization), it requests for full dump of key values. This causes flooding overhead when the network gets stable and most of key values are already synced up among all nodes.

In this diff we optimize the way KvStore requests for sync. Each node requests for KEY_DUMP by sending over it own hash info which is associated with {version, originator id, value} first, and after hash comparison its peer only respond for those key values of which hashes are different.

This optimizes kvstore full sync duration from `~55ms` to `~38ms` for a 32 * 32 topology

This implementation is compatible with older release of openr which doesn't support hash comparison.

Reviewed By: saifhhasan

Differential Revision: D6278161

fbshipit-source-id: 3012409cff7d5780a3d3014b27ee74429cf7d93c