Avoid watchdog crash when attaching to GDB
Summary:
When we attach process in GDB, all threads are jammed and resumes when we exit
from GDB. In this case, watchdog immediately finds all threads to be stuck and
fires off immediate crash (if gdb debugging time is > 5 minute, which is often
the case).

In this diff I am introducing guard against crashing process. Abort should be
triggered only when we see two consecutive runs failed. This should avoid us
from triggering crash everytime we attach process to GDB.

However if one attaches to GDB within 20s again then it is likely to crash :P
lol.

Reviewed By: plapukhov

Differential Revision: D6630611

fbshipit-source-id: 5091bb0232989016d62eb1f43edb6ce5405934f8