Partial Route Calculation in Decision
Summary:
Currently we do full SPF whenever Decision receives adjacency/prefix updates from KvStore. It slows down the debounce to program routes, and penalize twice for both ends of the SAME link.
Besides, when we prepare a graph for SPF calculation, we take both 'node' and its associated prefix as a vertex in graph, it also increases the number of vertices in graph and time complexity Djikstra is `O(E + VlogV)` :(

In this diff partial routing SPF is rolled out to achieve the following optimization:

1. When adjacency update is received, we enforce the bidirectional check and only trigger SPF re-calculation when there's actual link change. For example, for a graph like:

     n1: intf1 -------- n2: intf2

When we firstly receive adjacency update from n1 declaring that n2 is discovered at local interface intf1, we don't trigger SPF because n1 --- n2 is fully brought up yet. We only trigger it when n2 is also updating its adjacency with n1 discovered at n2's local interface intf2.

Similarly, when the link is brought down, we trigger SPF recalculation immediately when n1 is declaring n2 is no longer its neighbor, and we don't need to trigger n2 again when n2 is reporting missing of n1 few ms later as the link is no longer a bidirectionally validate one.

2. We also optimize the way we prepare a graph by mapping prefixes to node in order to shrink the size of graph. And when prefix update is received, we only re-calculate the routes based on our cached next-hop information instead of re-calculating SPF from scratch, as the graph itself actually don't change at all.

3. Backward compatibility is verified as the previous release of openr doesn't have remote interface field in Spark payload.

(Note: this ignores all push blocking failures!)

Reviewed By: saifhhasan

Differential Revision: D5978609

fbshipit-source-id: 76fa44bcf9e2ba2cf6f9120c95e17fa9aae7c2af