Make label distribution optional
Summary:
As of now none of our internal use cases uses label distribution. We anticipate
to use in near term future.

TG explicltly wants to reduce entries in KvStore and it seems that with this we
should be able to reduce #nodes keys in KvStore.

Further keep label distribution disabled by default.

Reviewed By: jstrizich, dombiv

Differential Revision: D6741954

fbshipit-source-id: fdc098e5ef4de8bf641a1c865ef7a691eecc23dd