Fix breeze decision path command
Summary:
This command started timing out recently because it can't talk to remote
hosts via their names (emulated node names) as they are not resolvable to IP
addresses.

Changes in this diff
- Use loopback address of a node instead of it's name to create client
- Fetch prefix-dbs, for loopback addresses, from KvStore. Decision lose
  information about prefix-type and hence it is hard to identify which prefixes
  to use and how to parse them.

p.s. There is one TODO in this diff, which doesn't need to be addressed now to
avoid code complexity. However it can be addressed easily once we move to
python3

Reviewed By: jstrizich

Differential Revision: D6814582

fbshipit-source-id: 7d2b570bbd2c3f5c97fb2ad260cde9876bb75ada