Get rid of local agent route to optimize Fib mem util
Summary:
This is a follow up task of D7053066. We want to avoid caching duplicated information in Fib (`routeDb_` and `agentRoutes_`) to lower mem util of Fib.

The way we do optimization is to maintain a `dirtyRouteDb_` flag inside Fib. Whenever an exception of route programming is caught by Fib, this flag is set to indicate what is cached currently may not be fully programmed successfully. So in the next iteration of either updateRoutes of syncRoutes, we enforce full sync between Fib and agent if we see this flag.

Reviewed By: saifhhasan

Differential Revision: D7102561

fbshipit-source-id: e302d6c9392dff1b715e9c153076b94ab2c1f16a