ROUTE_DB_GET in Decision mistakenly output duplicated paths upon same nexthop
Summary:
as titled.

Validation between Fib and Decision is currently failing in emulation with multiple pop node announcing same prefix. Similar issue observed in TG SJC that route validation to default route (::/0) is failing.

This is because when we query route database from decision, it doesn't deduplicate same path unfortunately. To give an example of the issue:

  node1 (ff80::1) ---<metric: 1>---node 2 (ff80::2) -----<metric: 1>-------node 3(::/0)
                                                     \
                                                       \
                                              < metric: 1 >
                                                          \
                                                 node 4 (::/0)

In this case when we query decision route database on node1, it will output 2 exactly the same nexthops {ff80::2, metric=2}. Fib is programming the correct nexthop though, the only thing affected is validation command `breeze fib validate`

The fix is to merge same nexthops when building routes in Decision

Reviewed By: saifhhasan, dombiv

Differential Revision: D7247824

fbshipit-source-id: 02cd891697e9188e83621f27ab4ba7f9ec2f773f