Fix glog command line arguments on a clean build
To support glog's command line arguments like --logtostderr, it requires gflags.

If gflags is not found, glog will still compile but the command line arguments will be absent.

As glog was created before gflags, that was the case on a clean build.
Note that on subsequent builds gflags already existed, so a second build 'magically' made the command line arguments appear.