run Dijkstra's shortest path without boost
Summary:
This diff aims to rework the decision module to no longer use boost's graph structs and Dijkstra implementation. I found when trying to implement a few new features in this module that the code is quite unapproachable. This has also led to some recent nasty bugs. The main reason for this is the boost structures are generic and therefore require a lot of ancillary processing to relate to our specific application - routing.

This new approach operates on a graph structure that is tightly related to our purpose. We maintain an adjacency list representation of the network topology where adjacencies are represented by the `Link` struct.

As outlined in the test plan, this change brings a slight regression to performance over the previous approach. In my view, this is outweighed by the wins. At this time, I feel code quality and readability are more important. If performance becomes an issue, I am confident we can squeeze more out of this.

Upside:
- Remove our dependency on boost.
- Simplify this critical module. The main source, Decision.cpp, is actually ~300 lines shorter now.
- enable simpler implementation of new features such as ordered fib.

Downside:
- large change
- slight perf regression

Reviewed By: saifhhasan

Differential Revision: D7529944

fbshipit-source-id: 322e9776e58274eaa11858c3dd06baf78221e6ae