Remove hardcoded protocol ID from NetlinkRouteSocket
Summary: This is the FIRST DIFF to completely remove constant protocol id in NetlinkRouteSocket. In this diff we just make protocolId to be a param of each API instead of a class member, so that it will support multiply protocols in future.  Underlying logic will be enhanced in the coming diff.

Reviewed By: jstrizich, saifhhasan, hanli0612

Differential Revision: D8516149

fbshipit-source-id: cc5b7c4778222e36c0c8ad66e3731994ea8a0a1e