log build information at the beginning
Summary:
On rsws, I observed that build information gets logged just before openr is
stopped. This happens because we log build information in stdout which gets
flushed only when process is about to exit.

This was confusing as if build information corresponds to previously running
version of new version that was started.

Using `LOG(INFO)` stream to guarantee the order on log output.

Reviewed By: smaheshm

Differential Revision: D8654785

fbshipit-source-id: 7d27c6d42073ff5f57ec809aecb5169b7dba8132