fix graceful shutdown
Summary:
=== What
When build in dev mode, shutting down open/r raises errors on the lines of
`heap memory being accessed after being released`. This is mostly the issue with
order of destruction of variables.

This issues shows up in emulation with dev build when services are restarted.
Fixing the issue with Fib/System handler and their thrift-server.

=== Why
Having clean shutdown is important, especially when we expect a certain config
update needs to be flushed out to disk (node/link overload) before shutting
down. This heap memory access only occurs during shutdown and in newer compiler
version it can be of an issue if heap that is freed gets allocated to some new
variable

Reviewed By: cenzhao

Differential Revision: D8670583

fbshipit-source-id: 7a845bf591ca119f2e1518be0d6341cc26af3103