Bug fix in duplicate socket RID
Summary:
When one of the parallel link goes down between a pair of neighbor, Open/R
crashes because of duplicate socket RID which is only built using neighbor's
name.

Now using monotonically increasing counter to frame peer-socket-id which is
going to be unique for a particular instance of Open/R. This will solve problem
to it's root.

Reviewed By: GirasoleY

Differential Revision: D9122789

fbshipit-source-id: f305082f142e0ad44b82a60d790fcb6a8efffe9f