Update adjacency DB when the neighbors v4 address changes after forming adjacency
Summary:
SOMA requires openr to allocate link local v4 addresses. When a new IP is re-assigned after duplicate address detection, the i/f and spark hello pkts reflect the new IP, but the adj DB is not updated.

The fix is to consider neighbor restart when the v4 address change is deteted.

Reviewed By: cenzhao, ammubhave

Differential Revision: D9272769

fbshipit-source-id: 7efef25aca8110c32ab7d55f15bf248485e1f462