Fix flaky unit test
Summary:
Unit test uses dumpAll API to check if store is empty. After adding changes for accurate TTL expirey, KEY_DUMP returns the keys only if the remaining TTL is above a configured threshold. So it is likely that dumpAll returns empty even if there are some keys left with TTL below the threshold TTL.

Fix is to add a new command type to return all keys regardless of the remaining TTL value

Reviewed By: cenzhao

Differential Revision: D9299959

fbshipit-source-id: a55284d12df1b3768316fd6a73800c75bf8c3cad