fix bug where nlSocket goes out of scope before handlers are constructed
Summary:
the nlSocket is declared within an if block so if main thread executes past here before handlers copy construct the socket pointer the socket is destructed. Solution is to simply capture the smart pointer by value for the threads

This is causing fatals on all ci jobs.

Next step is to fix the call in ~NlSocket that relies on being in the eventLoop

Reviewed By: cenzhao, hanli0612

Differential Revision: D9687199

fbshipit-source-id: 4e3e0836627ecc9f1cf28cff97fc877395d2280d