ensure new prefix always gets synced in prefix-manager
Summary:
We pre-maturely return if prefix to be advertised exists on interface and we
don't end up announcing it via PrefixManager.

Do both these steps independently to ensure prefix is always advertsed in Prefix
Database.

p.s. Observed issue in emulation where netlink socket call fails and potentially
leads to a state where prefix is assigned on interface but not advertised. This
diff fixes it.

Reviewed By: smaheshm

Differential Revision: D10382119

fbshipit-source-id: 86a9332cfd6d7ebcf8345a588e3b35cdbcfec800