Make callbacks noexcept
Summary:
This is a move towards enforcing proper use of callbacks. An exception can never
be thrown inside callback as it can affect internal state of netlink sockets
(and hence ref-couting problem). By setting noexcept callback user is enforced
to not to do so.

Fixed a bug where in ROUTE callback `isValid` attribute was not set. This
in-turn fixed a broken UT.

Reviewed By: jstrizich

Differential Revision: D10513608

fbshipit-source-id: b178c183e489d576e092d62d697e5a11c0866652