add superclass for openr modules
Summary:
The aim here is to hoist certain common aspects of openr modules into an abstract parent class, `OpenrEventLoop`. The immediate gain from this is to reduce the number of TCP ports we connect to by eliminating tcp command ports for modules that only bind the port for the benefit of breeze (eg. Decision, Fib, LinkMonitor). We can continue interacting with these modules remotely via a single thrift service that handles demultiplexing commands to the appropriate module. see D6904142.

For now, the superclass does the following:
- inherits from ZmqEventLoop
- standardizes socket options, types and urls per module.
- removes a lot of repetitive code from the modules.

Going forward, this allows us to more easily make changes to our architecture. For example:
- it would be much easier to change our thread communication mechanism to something like a message queue.
- can come up with a standard mechanism for modules to publish and subscribe built on top of a single underlying transport
- we now have a place to drop anything that is standard across modules (e.g. counter collect and publication could also get hoisted up here)

Reviewed By: saifhhasan

Differential Revision: D9114687

fbshipit-source-id: a3ab71bff56af40ae332cece65e5cbe3c19e22d0