Remove redundant flag from openr config
Summary:
`authenticate_peer_common_name` has been tightly coupled with
`enable_secure_thrift_server` ... If later one is set then former one must be
set and vice versa.

In this diff I am removing the flag as it practically servers no purpose as of
now for our usecase. I encountered setting this flag in emulation (while
disabling secure thrift and thought of removing it). LMK if this is not
approrpiate and we might have usecase where we won't authenticate-peer-name but
secure-thrift is enabled

Reviewed By: jstrizich

Differential Revision: D13705320

fbshipit-source-id: 02437eaabfa44fb26c93e3feb98b7a887e02f052