make client socket a class member
Summary:
We need  to close the transport on any thrift socket we initialize outside a `with` statement. This reworks OpenrClient logic a bit to decide which transport method to use once and for all in `__init__` and take care to close any unneeded or failed transports.

In the future we will have one transport, secure thrift, and ought to initialize any client in a `with` statement. For now, this seems the best way to have graceful fallback to legacy transports.

Fixes T39234153

Reviewed By: saifhhasan

Differential Revision: D13857720

fbshipit-source-id: 62cbd493e46b682d082e4e521aecf6b79d88b593