thrift definitions for MPLS route and APIs
Summary: Adding MPLS structs and route APIs. This is going to be duplicated in `fboss/agent/if/ctrl.thrift`. Finalizing on schema between fboss and open/r is important for further work. I am trying to look into re-using code but fb303 service is making it difficulat (as open/r doesn't inherit fb303 service as of now).

Reviewed By: phshaikh

Differential Revision: D13915092

fbshipit-source-id: 14b3ec85db320a3da284a709b6275a98c2911a50