Add forwarding type attribute to advertised prefixes
Summary:
With introduction of SR we will have two kind of forwarding.
1. IP (current and default) => Where packet is looked up for prefix on every hop
   and forwarded via nexthop information
2. MPLS (new) => packet is inspected up by prefix on very first hop, label
   (label of node advertising prefix) and within the network packet is looked
   up by it's top label until it reaches the destination

In order to attract traffic via IP forwarding of MPLS we need a way to tell
Open/R which route to install for the prefix. This can be achieved by adding
`forwardingType` attribute to PrefixEntry.

Next Step - Modify decision to look at this attribute and create IP->MPLS route
for prefixes with `forwardingType == MPLS`

Reviewed By: smaheshm

Differential Revision: D14115914

fbshipit-source-id: 43f9aeb987d05ab41a366355ff59149d99353a16