Re-organize prefix-entries in Decision module
Summary:
As we introduce MPLS routes and other protocol route re-distribution, it is
important to organize prefixes along with their attributes (not just the node
name).

This diff simply moves PrefixEntry object from `prefixDatabases_` to `prefixes_`
and refactor the associated code.

One major change - If any attribute of PrefixEntry object changes then we will
consider it as an update which earlier code didn't consider.

Reviewed By: jstrizich

Differential Revision: D14251866

fbshipit-source-id: 0e8c68632374c176ff9bfd986ab2cb7f3abd0f16