Add mpls routes for node labels
Summary:
For each node label we will create MPLS route that matches the packet destined
to node (with nodeLabel) and forward them towards the node over the shortest
path. LFAs will also be accounted for.

In short nexthops for a nodeLabel will be very similar to nexthops for IpPrefix
advertised by that node. There are subtle differences
- For own nodeLabel we program a `POP_AND_LOOPKUP` route (no nexthop address specified)
- For other nodeLabel we find all nexthops (including LFA) towards the node.
  Then assign `PHP` with a nexthop if node it is destination or else
  assign `SWAP` (with dst node label) action. (all these nexthops are valid for non loop free routing).

NOTE: Fib will take nexthops and program only shortest cost nexthop. In addition
for MplsRoute FIB will prefer to program nexthops with same action with
preference to PHP (forward to directly connected node)

Reviewed By: jstrizich

Differential Revision: D14294040

fbshipit-source-id: 703d6e6a2728a120e302d41e7f5d73d6a367bac1