Ensure withdraw prefix is client or type specific
Summary:
IpPrefix + type/client uniquely identifies a instance to withdraw. Even though thrift interface takes PrefixEntry, PrefixManagerClient, is only taking IpPrefix as input, this causes conflict as to which specific prefix to be deleted. Going forward as we enable multi client support etc we need to uniquely delete a specific instance of prefix mgr entry.

Ensured that PrefixEntry is deleted only if type matches.
Ensured that all the prefixes passed for withdraw must exist, i.e. we only act on the message, if full message can be acted upon without any errors.
Another reason for doing this change is, we can use prefix type in input parameters to figure out if we need to update persistent store or not (for next diff) without checking values in prefixMap_.

Reviewed By: saifhhasan

Differential Revision: D14453541

fbshipit-source-id: da3e3d45a56c29f7025a6fcd1396e75d91300aa2