Support ephemeral prefix manager
Summary:
Support ephemeral prefix entries.

To support high scale for prefix manager when large number of bgp routes at fsw are redistributed into openR, we need ephemeral PrefixEntry so that these entries and their updates will not be written to file. BGP will ensure that all these prefixes are updated before fib sync (3*KA) for GR support.

Added optional field in PrefixEntry to indicate ephemeral/persistent. Cli will display the same.

Split the persistPrefixDb into persistPrefixDb, updateKvStore so we do one or both depending on the type of prefix which is being updated.

Ensured that persistent store with only persistent entries is updated only when needed.

Fixed to avoid LOAD command from saving to file, as it's read operation.

Introduced configurable backoff times for persistent store, this is used to avoid sleep in tests. We can configure to zero and expect it to work inline.

Added a counter to keep track of number of writes to disk. We can export it if we desire.

Reviewed By: saifhhasan

Differential Revision: D14453741

fbshipit-source-id: 73cc36541952207ac5c52d94e0fb5cf13fbdb2e7