Change include linux/if.h to net/if.h
Summary:
Rest of openr uses net/if.h, except fbnl.
This is causing conflicts when including both fbnl, and other modules together
in the same file.
The reason is both linux/if.h and net/if.h declare a lot of consts and structs
which are common. I am changing linux -> net in fbnl as well so that it is
consistent with rest of openr and doesn't cause conflicts when compiling
together.

Reviewed By: cenzhao

Differential Revision: D14671849

fbshipit-source-id: 2c315e7c8ad54260cd6f440e48b692da544d2fc0