Create system buildscript generator based on fbcode_builder logic (#41)
Summary:
Title: [fbcode_builder] Create system buildscript generator based on fbcode_builder logic

Description:
As discussed in #38 the build script `build_openr.sh` is broken, mainly due to dependency issues. I suggested in #40 to use the fbcode_builder logic to create the script with the same dependencies of the build logic used by the CI system.

The debian_system_builder allows to create a build script that can be run directly on the host similar to the shell_builder. In addition, the shell_builder is adjusted for a system build/ install.
Pull Request resolved: https://github.com/facebook/openr/pull/41

Differential Revision: D14819285

Pulled By: saifhhasan

fbshipit-source-id: b16114432faee46a28fadf1e0fa1e891b8748010