Remove unnecessary part of hwmp routing
Summary:
Thought about this a bit, we do not need most of the HWMP mechanisms
for our use cases. Since we don't have a way of forming reactive routes, there
isn't any use for the rest of the code.
Nix it, it will always be in the history if we need to bring parts of this
back.

I have replaced the actual routing algorithm that gets used with a custom DV
routing algorithm with ideas taken from HWMP, and this algorithm seems to be
working.

Reviewed By: mariof

Differential Revision: D14941022

fbshipit-source-id: 43a11cfdc5f4f2601f18e5c2012f42df039ef952