Add Metric vector in prefix entry.
Summary:
Adding metric vector optional field in prefix entry.
Default constructor which thrift generates, doesn't have a way to pass optional fields and retain them as optional, instead we have to pass some value.
Created a utility method and used it every where so that optional fields are not set if not needed.

Reviewed By: jstrizich

Differential Revision: D15036665

fbshipit-source-id: ad7e1339be713b71ef93bd09e0046b0cac556584