add function to compare MetricVectors
Summary:
Nanda, let me know if this was what you had in mind for comparing. There are a few edge cases I'm not sure about and marked with `TODO`

Nanda. Saif: I'm concerned that the metric vector structure is too complex and that more of the BGP preference could be encoded into a simpler structure. What do you think? We can discuss offline

Differential Revision: D15053607

fbshipit-source-id: adcce7d740ec46120179eff9c32a350632ab41bd