Support multiple addresses in GatewayConnectivityMonitor
Summary:
Currently `GatewayConnectivityMonitor` accepts a single IP and port to use to assess whether a device has WAN connectivity and can be used as a gateway.

During S177019, we observed that this could potentially cause network instability in the case that the network and backhaul are perfectly fine, but the remote host selected for the gateway monitor is down (e.g. if using a Facebook IP for this, we don't want a partial Facebook outage to take down an unrelated mesh network).

In order to improve reliability, add support for checking multiple IPs to determine if a device is a gateway. If multiple unrelated hosts are selected, this change will prevent `GatewayConnectivityMonitor` from having a single point of failure.

Reviewed By: ammubhave

Differential Revision: D15217499

fbshipit-source-id: 8b2fe34eb21833f22cb7df1ff5ef13afc1a0b065