Add 1.1.1.1 to gateway connectivity monitor addresses
Summary:
During S177019, we observed that this could potentially cause network instability in the case that the network and backhaul are perfectly fine, but the remote host selected for the gateway monitor is down (e.g. if using a Facebook IP for this, we don't want a partial Facebook outage to take down an unrelated mesh network).

To help mitigate this in the short term, add the Cloudflare DNS IP to the addresses monitored by `GatewayConnectivityMonitor`.

Reviewed By: ammubhave

Differential Revision: D15217813

fbshipit-source-id: 71b8e52ba81acb55445960d25e9e8b71a1c977f1