Remove duplicated default values; configure all via magma
Summary:
Currently we use a set of config files to set up `fbmeshd`, including configurations set from several different locations (see D14143939 for details):
0. `fbmeshd` C++ gflags defaults
0. `run_fbmeshd.sh`
0. `/etc/fbmeshd.conf`
0. `/etc/fbmeshd.conf.d/*`

This isn't ideal for several reasons:
0. Both 1 and 2 above set defaults, which don't always match
0. Adding a new flag in C++ requires adding it to `run_fbmeshd.sh` before it can be used
0. Making individual flags magma-configurable requires many similar changes to `/etc/fbmeshd.conf`

In order to improve this, we consolidate the logic to still pull configs from `/etc/fbmeshd.conf` and `/etc/fbmeshd.conf.d/*`, but instead of explicitly defining the variable names, we now search for all variables of the format `CONFIG_fbmeshd_*`. This allows us to avoid all of the above problems:
0. We now only get defaults from the C++ (apart from specific SoMA-only overrides in `/etc/fbmeshd.conf`)
0. New flags are automatically discovered
0. All flags are immediately magma-configurable (still via code in `/etc/fbmeshd.conf`).

Additionally, with this change, we are able to remove most of the config overlays as they are now configurable using magma.

To avoid any impact on users, the `fbmeshd` configs from the overlays have been transferred into the magma properties for meshes as appropriate.

**Note:** We still have to keep the `habari`, `surf`/`surf2`, and `cellc` overlays around as they have a separate admin password. To do this, the overlay folder is kept with a new `.<overlay>` file in it to allow us to keep an "empty" overlay folder in mercurial.

Reviewed By: mariof, ammubhave

Differential Revision: D15302642

fbshipit-source-id: b689e65b66e969c917624c54c5b751c3baba2b84