Fix run_fbmeshd.sh for passwords
Summary:
In D15302642 I made a change that incorrectly interpreted the output of `set` as valid flags for `fbmeshd`. They were not, since they could include single-quotes around strings, which would then get passed to `fbmeshd` as though it was part of the user-specified input string.

This fixes that by instead generating the `fbmeshd` arg array using the bash-provided values of those variables.

Reviewed By: bcopeland

Differential Revision: D15357925

fbshipit-source-id: 840c443369391dc1f433b693554e277e3d90036a