Benchmark test for the persistent store
Summary:
This piece of code is to perform the benchmark test for the persistent store, including #reads/second, #writes/second and #create/destroy per second.

(10, 100, 1000, 10000) random (key, value) pairs are generated as inputs.

Three benchmark functions are included in the file PersistentStoreBenchmark.cpp, which are: BM_PersistentStoreWrite(.), BM_PersistentStoreLoad(.) and BM_PersistentStoreCreateDestroy(.).

Reviewed By: saifhhasan

Differential Revision: D15532375

fbshipit-source-id: 49526cb178ded7643cf5f70805256ed14b9a9205