Attr to distinguish between LFA vs 2nd-shortest path
Summary:
Fib module forwards only smallest metric nexthops to FibAgent. This is inteded
to not program LFA routes and keep them in backup.

However in case of KSP2 we want both shortest and non-shortest paths to be
programmed. Adding `useNonShortestRoute` attribute to nextHop to achieve the
same. Nexthops that have this attribute set will always be programmed

Reviewed By: jstrizich

Differential Revision: D15728459

fbshipit-source-id: 46170e55ce8906fc3d73a44d3a646535f6ae319a