Set NetlinkProtocolSocket callbacks after libnl cache initialization
Summary:
NetlinkProtocolSocket callbacks will only be called after all NetlinkSocket caches have been initialized. Before, we were seeing SIGSEGV once in a while because the linkCache was not instantiated before it was used in the callback.

Note - As we transition from libnl to our Netlink implementation, we will get rid of the libnl caches in the future.

Differential Revision: D16080400

fbshipit-source-id: 21541adc8ac8b95677c1af89cffed08bf15025f7