Add sync routes thread to watchdog
Summary:
The idea behind D15632816 was to crash fbmeshd when it gets into a state where the syncRoutes loop gets stuck. That diff didn't quite work out and was backed out in D15936840.

This diff adds the syncRoutes event loop to the watchdog, so fbmeshd is killed if it gets stuck. I had to refactor a few things to make it work.

Reviewed By: bcopeland

Differential Revision: D16114209

fbshipit-source-id: 1874c0f2cb48bac9087f874da8c06d2ab7240408