Combine the bitrate based metric with RSSI based metric
Summary:
We sometimes see weird bitrates on poor links. This is an attempt to weigh RSSI based metric into the overall metric. The weight is parametrized, so we can turn it off (or use only RSSI).

One interesting thing here is that the RSSI value is in "the wrong direction". If rssi is not symmetric, this combines rx rssi with tx bitrate. Not a big deal, or maybe that's even better.

Reviewed By: jcard0na

Differential Revision: D16153852

fbshipit-source-id: f14de0c6c4a6f5142bef35cd763b9ebc7e293e4f