Replace ipc socket used by PersistentStore to inproc socket (12/N)
Summary:
To give a little bit background here, right now. OpenrEventLoop class is opening 3 types of sockets based on the constructor parameter we passed in from its child class( e.g. Fib, Decision, etc. )

- InProc socket
- Tcp sockect
- Ipc socket

As the efforts to remove zmq socket usage inside Open/R and migrating to thrift, we will remove tcp and ipc socket serving as the purpose.

For IPC socket usage, PersistentStore is the ONLY usage of it. ( Well, this is NOT exactly true since we still have forceCrashUrl combined to IPC socket in Open/R. But it is NOT instantiated through OpenrEventLoop class ). And persistentStore is ONLY used inside Open/R module. So it can completely be replaced with INPROC socket.

By removing PersistentStore usage of IPC socket, we can remove the option inside OpenrEventLoop for "ipcUrl" and codebase to open IPC socket. This will come later separately once this change is done.

Reviewed By: saifhhasan

Differential Revision: D16272999

fbshipit-source-id: 63214735ef6a1f487292f94010857f699ef76101