Fix valid timeoutId being treated as invalid
Summary:
`authsae` reserves the timer ID `0` for internal usage. Although `ZmqEventLoop::scheduleTimeout()` can return this timeout ID, we did not run into this edge case because we were scheduling a lot of other timers as well.

To resolve this, adjust our timeout ID abstraction in `AuthsaeCallbackHelpers` to enforce that the timeout ID will never be `0`. We do this by applying the transform

```
publicTimeoutId = (privateTimeoutId + 1)
```

Reviewed By: ammubhave

Differential Revision: D16356964

fbshipit-source-id: 6efff3f122115714f7133c99ebb5ff169ac9b618